+++ macros/src/lib.rs
@@ -9,7 +9,6 @@
 extern crate proc_macro;
 
 mod int;
-mod itm_update_prescaler;
 mod sv;
 mod vtable;
 
@@ -18,11 +17,6 @@ use proc_macro::TokenStream;
 #[proc_macro]
 pub fn int(input: TokenStream) -> TokenStream {
     int::proc_macro(input)
-}
-
-#[proc_macro]
-pub fn itm_update_prescaler(input: TokenStream) -> TokenStream {
-    itm_update_prescaler::proc_macro(input)
 }
 
 #[proc_macro]