+++ src/itm/mod.rs
@@ -8,11 +8,6 @@
 
 mod port;
 
-/// Updates the SWO prescaler register to match the baud-rate defined at
-/// `Drone.toml`.
-#[doc(inline)]
-pub use drone_cortex_m_macros::itm_update_prescaler as update_prescaler;
-
 pub use self::port::Port;
 
 use crate::{
@@ -35,8 +30,8 @@ pub fn is_enabled() -> bool {
     unsafe { read_volatile(ITM_TCR as *const u32) & 1 != 0 }
 }
 
-/// Returns `true` if the debug probe is connected and listening to the ITM port
-/// `port` output.
+/// Returns `true` if the debug probe is connected and listening to the output
+/// of ITM port number `port`.
 #[inline]
 pub fn is_port_enabled(port: usize) -> bool {
     #[cfg(feature = "std")]
@@ -83,13 +78,40 @@ pub fn flush() {
 /// Generates an ITM synchronization packet.
 #[inline]
 pub fn sync() {
+    #[cfg(feature = "std")]
+    return;
     let mut cyccnt = unsafe { dwt::Cyccnt::<Urt>::take() };
     cyccnt.store(|r| r.write_cyccnt(0xFFFF_FFFF));
 }
 
 /// Updates the SWO prescaler register.
+///
+/// # Examples
+///
+/// ```
+/// # #![feature(proc_macro_hygiene)]
+/// # drone_core::config_override! { "
+/// # [memory]
+/// # flash = { size = \"128K\", origin = 0x08000000 }
+/// # ram = { size = \"20K\", origin = 0x20000000 }
+/// # [heap]
+/// # size = \"0\"
+/// # pools = []
+/// # [probe]
+/// # gdb-client-command = \"gdb-multiarch\"
+/// # [probe.swo]
+/// # reset-freq = 8000000
+/// # baud-rate = 115200
+/// # " }
+/// use drone_core::log;
+/// use drone_cortex_m::itm;
+///
+/// itm::update_prescaler(72_000_000 / log::baud_rate!() - 1);
+/// ```
 #[inline]
 pub fn update_prescaler(swoscaler: u32) {
+    #[cfg(feature = "std")]
+    return;
     let mut acpr = unsafe { tpiu::Acpr::<Urt>::take() };
     acpr.store(|r| r.write_swoscaler(swoscaler));
     sync();
@@ -97,52 +119,61 @@ pub fn update_prescaler(swoscaler: u32) {
 
 #[doc(hidden)]
 #[macro_export]
-macro_rules! itm_init {
+macro_rules! itm_set_log {
     () => {
-        $crate::reg::assert_taken!(dwt_cyccnt);
-        $crate::reg::assert_taken!(itm_tpr);
-        $crate::reg::assert_taken!(itm_tcr);
-        $crate::reg::assert_taken!(itm_lar);
-        $crate::reg::assert_taken!(tpiu_acpr);
-        $crate::reg::assert_taken!(tpiu_sppr);
-        $crate::reg::assert_taken!(tpiu_ffcr);
+        const _: () = {
+            $crate::reg::assert_taken!("dwt_cyccnt");
+            $crate::reg::assert_taken!("itm_tpr");
+            $crate::reg::assert_taken!("itm_tcr");
+            $crate::reg::assert_taken!("itm_lar");
+            $crate::reg::assert_taken!("tpiu_acpr");
+            $crate::reg::assert_taken!("tpiu_sppr");
+            $crate::reg::assert_taken!("tpiu_ffcr");
 
-        #[no_mangle]
-        extern "C" fn drone_log_is_port_enabled(port: u8) -> bool {
-            $crate::itm::is_port_enabled(port as usize)
-        }
-
-        #[no_mangle]
-        extern "C" fn drone_log_port_write_bytes(
-            port: u8,
-            exclusive: bool,
-            buffer: *const u8,
-            count: usize,
-        ) {
-            let bytes = unsafe { ::core::slice::from_raw_parts(buffer, count) };
-            if exclusive {
-                $crate::itm::write_bytes_exclusive(port, bytes);
-            } else {
-                $crate::itm::write_bytes(port, bytes);
+            #[no_mangle]
+            extern "C" fn drone_log_is_enabled(port: u8) -> bool {
+                $crate::itm::is_port_enabled(port as usize)
             }
-        }
 
-        #[no_mangle]
-        extern "C" fn drone_log_flush() {
-            $crate::itm::flush();
-        }
+            #[no_mangle]
+            extern "C" fn drone_log_write_bytes(
+                port: u8,
+                exclusive: bool,
+                buffer: *const u8,
+                count: usize,
+            ) {
+                let bytes = unsafe { ::core::slice::from_raw_parts(buffer, count) };
+                if exclusive {
+                    $crate::itm::write_bytes_exclusive(port, bytes);
+                } else {
+                    $crate::itm::write_bytes(port, bytes);
+                }
+            }
+
+            #[no_mangle]
+            extern "C" fn drone_log_flush() {
+                $crate::itm::flush();
+            }
+        };
     };
 }
 
-/// Initializes ITM logging.
+/// Sets ITM as default logger.
 ///
 /// # Examples
 ///
 /// ```
-/// #![feature(proc_macro_hygiene)]
-/// use drone_cortex_m::itm;
+/// # #![feature(proc_macro_hygiene)]
+/// use drone_cortex_m::{cortex_m_reg_tokens, itm};
 ///
-/// itm::init!();
+/// cortex_m_reg_tokens! {
+///     struct Regs;
+///     !dwt_cyccnt;
+///     !itm_tpr; !itm_tcr; !itm_lar;
+///     !tpiu_acpr; !tpiu_sppr; !tpiu_ffcr;
+/// }
+///
+/// itm::set_log!();
 /// ```
 #[doc(inline)]
-pub use crate::itm_init as init;
+pub use crate::itm_set_log as set_log;